perf(web): optimize LCP by splitting hero to server component and deferring notification init
- Convert HomeHero to async Server Component so h1#hero-heading renders during SSR
- Extract HeroCardStack as thin Client Component with Suspense boundary
- Remove useEffect from NotificationBell; fetch unread count on hover/focus only
- Add fetchUnreadCount store action for lightweight badge fetch
- Lazy-load NotificationBell via next/dynamic with ssr:false
- Memoize NotificationPopover and NotificationRow
- Split animations.scss: 521-line critical path vs 992-line page styles
- Remove unused styledComponents compiler option from next.config.ts
- Raise tsconfig target to ES2020 (browserslist already requires modern browsers)
- Fix RouteChangeAnnouncer hydration mismatch with suppressHydrationWarning
